+++ software/firmware/lib/src/iep.c
@@ -11,7 +11,7 @@
 volatile register uint32_t __R31;
 #endif
 
-void iep_set_us(uint32_t value)
+void iep_set_us(unsigned int value)
 {
 	iep_stop();
 	iep_reset();
@@ -39,7 +39,7 @@ void iep_stop()
 }
 
 
-inline int8_t iep_check_evt_cmp(uint32_t compare_channel)
+inline int iep_check_evt_cmp(unsigned int compare_channel)
 {
 	if(CT_IEP.TMR_CMP_STS & (1 << compare_channel))
 		return 0;
@@ -47,31 +47,31 @@ inline int8_t iep_check_evt_cmp(uint32_t compare_channel)
 		return -1;
 }
 
-inline void iep_clear_evt_cmp(uint32_t compare_channel)
+inline void iep_clear_evt_cmp(unsigned int compare_channel)
 {
 	CT_IEP.TMR_CMP_STS |= (1 << compare_channel);
 }
 
-void iep_enable_evt_cmp(uint32_t compare_channel)
+void iep_enable_evt_cmp(unsigned int compare_channel)
 {
 	CT_IEP.TMR_CMP_CFG_bit.CMP_EN |= (1 << compare_channel);
 }
 
-void iep_disable_evt_cmp(uint32_t compare_channel)
+void iep_disable_evt_cmp(unsigned int compare_channel)
 {
 	CT_IEP.TMR_CMP_CFG_bit.CMP_EN &= ~(1 << compare_channel);
 }
 
-void iep_set_cmp_val(uint32_t compare_channel, uint32_t value)
+void iep_set_cmp_val(unsigned int compare_channel, unsigned int value)
 {
 	/* Hack to address the CMPN registers (offset=0x48 from CT_IEP) */
-	*((uint32_t *) &CT_IEP + 18 + compare_channel) = value;
+	*((unsigned int *) &CT_IEP + 18 + compare_channel) = value;
 }
 
-uint32_t iep_get_cmp_val(uint32_t compare_channel)
+unsigned int iep_get_cmp_val(unsigned int compare_channel)
 {
 	/* Hack to address the CMPN registers (offset=0x48 from CT_IEP) */
-	return *((uint32_t *) &CT_IEP + 18 + compare_channel);
+	return *((unsigned int *) &CT_IEP + 18 + compare_channel);
 }
 
 void iep_init()
